04-Advanced-TensorFlow/02-Mobile-ML-Applications/fashion-mnist-tflite/requirements.txt
tensorflow>=2.9.1
tensorflow-datasets>=4.6.0
numpy>=1.21.0
matplotlib>=3.4.0
Pillow>=8.0.0
absl-py>=0.12.0 